requirements.txt
requests
load_dotenv
beautifulsoup4